First commit for organized dev branch in multiple files (Instead of everything inside the main function).
In this commit, the code is running without problems (Only obstacle avoidance - algo issue)
